Updated to use margin property
Margin allows for much more consistent design compared to padding

Margin essentially compresses the element within their boundary to fit the container

While padding on the other hand pushes the container outward to give the elements the required spacing

With margin, since the size of the container remains constant you dont have to do any math to do pixel perfect design

To use margin the mental model should be to first figure out the container in which all the elements are going to be placed.

Then put the all the elements in a wrapper class, then use margin on the wrapper class.
Also any kind of layout should also be done at the wrapper class and not at container level to avoid any kind of unnecessary inheritance