Done using padding version
Essentially the real crux that need to be remembered from here is the power of wrapper class

previously padding wasn't working as intended because in css everything is a box as such any effect that is applied at any level will only effect itself and its immediate self.

This is why using a wrapper class is super important because it leads to having a single container to have all the elements.

But I guess it all depends on the type of effect you want to achieve